Change test filename
